Multipagejson (#2)
* Add multipage support first step, i.e. single thread
The target being to process json in another thread and execute new query
Or do multiple query in multiple thread.

* Add multipage support. For loop that iterates until data field in json
is empty. Maybe we can enhance this part.
Refacto the regular expression part in Dicitionnary.cpp so that we
create it once and copy it in both common and uncommon processing
thread.